refactor: Split config methods from BuiltInPlugin
