+++ src/main/java/io/github/null2264/cobblegen/integration/BuiltInPlugin.java
@@ -1,7 +1,5 @@
 package io.github.null2264.cobblegen.integration;
 
-import blue.endless.jankson.*;
-import com.google.gson.Gson;
 import io.github.null2264.cobblegen.CGPlugin;
 import io.github.null2264.cobblegen.CobbleGenPlugin;
 import io.github.null2264.cobblegen.compat.Loader;
@@ -24,8 +22,6 @@ import net.minecraft.world.level.material.Fluids;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -34,6 +30,7 @@ import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static io.github.null2264.cobblegen.CobbleGen.MOD_ID;
+import static io.github.null2264.cobblegen.config.ConfigHelper.loadConfig;
 import static io.github.null2264.cobblegen.util.Util.notNullOr;
 
 @CGPlugin
@@ -41,8 +38,6 @@ public class BuiltInPlugin implements CobbleGenPlugin
 {
     private static final Path configPath = Loader.getConfigDir();
     private static final File configFile = new File(configPath + File.separator + MOD_ID + ".json5");
-    private static final Jankson jankson = Jankson.builder().build();
-    private static final Gson gson = new Gson();
     @Nullable
     private static ConfigData config = null;
 
@@ -59,7 +54,7 @@ public class BuiltInPlugin implements CobbleGenPlugin
     @Override
     public void registerInteraction(CGRegistry registry) {
         CGLog.info((!isReload ? "L" : "Rel") + "oading config...");
-        if (config == null || isReload) config = loadConfig(isReload);
+        if (config == null || isReload) config = loadConfig(isReload, configFile, config);
         if (config == null) throw new RuntimeException("How?");
 
         AtomicInteger count = new AtomicInteger();
@@ -125,64 +120,5 @@ public class BuiltInPlugin implements CobbleGenPlugin
     public void onReload() {
         CGLog.info("Reloading built-in plugin...");
         isReload = true;
-    }
-
-    /**
-     * @deprecated Removed when Jankson released their proper null filter
-     */
-    @Deprecated
-    @Nullable
-    private static JsonElement filter(JsonElement json) {
-        JsonElement result = null;
-        if (json instanceof JsonObject finalResult) {
-            finalResult.keySet().forEach(key -> {
-                JsonElement element = finalResult.get(key);
-                if (!(element instanceof JsonNull) && element != null) filter(element);
-                else finalResult.remove(key);
-            });
-            result = finalResult;
-        } else if (json instanceof JsonArray finalResult) {
-            finalResult.forEach(element -> {
-                if (element instanceof JsonObject) filter(element);
-            });
-            result = finalResult;
-        }
-        return result;
-    }
-
-    private static ConfigData loadConfig(boolean reload) {
-        String string = reload ? "reload" : "load";
-        try {
-            CGLog.info("Trying to " + string + " config file...");
-            JsonObject json = jankson.load(configFile);
-            return gson.fromJson(json.toJson(JsonGrammar.COMPACT), ConfigData.class);
-        } catch (Exception e) {
-            CGLog.error("There was an error while " + string + "ing the config file!\n" + e);
-
-            if (reload) {
-                CGLog.warn("Falling back to previously working config...");
-                return config;
-            }
-
-            val newConfig = ConfigData.defaultConfig();
-            if (!configFile.exists()) {
-                saveConfig(newConfig);
-            }
-            CGLog.warn("Falling back to default config...");
-            return newConfig;
-        }
-    }
-
-    private static void saveConfig(ConfigData config) {
-        try {
-            CGLog.info("Trying to create config file...");
-            FileWriter fw = new FileWriter(configFile);
-            JsonElement jsonElement = Jankson.builder().build().toJson(config);
-            JsonElement filteredElement = filter(jsonElement);
-            fw.write((filteredElement != null ? filteredElement : jsonElement).toJson(JsonGrammar.JSON5));
-            fw.close();
-        } catch (IOException e) {
-            CGLog.error("There was an error while creating the config file!\n" + e);
-        }
     }
 }
